Move OnlyWhenRunningOnTheJvmCondition
